+++ CHANGELOG.md
@@ -28,6 +28,7 @@ This is used for recording Invader's changes. This changelog is based on
   tuned for more changes.
 - invader-build: Greatly improved the accuracy of firing positions so everything
   (or nearly everything) firing position related works
+- invader-build: Fixed some AI conversations not playing
 - invader-edit-qt: Fixed tag count not being updated on the main window if the
   tags directory is empty
 - invader-edit-qt: Fixed an unnecessary warning appearing on Windows when